refactor: fix meteor unit tests
