+++ .gitignore
@@ -10,6 +10,7 @@ do_not_commit
 public/fonts/icons*
 public/fonts/brand*
 public/fonts/outline*
+/botfront/client/fonts/
 botfront/cypress/videos
 botfront/cypress/screenshots
 bf_project_id.txt
