+++ botfront/.storybook/config.js
@@ -1,5 +1,6 @@
 import { configure } from '@storybook/react';
 import 'semantic-ui-css/semantic.min.css';
+import '../client/main.less';
 
 function loadStories() {
     const req = require.context('../stories', true, /\.stories\.js$/);
