+++ botfront/client/main.less
@@ -1,9 +1,23 @@
 @import url('https://fonts.googleapis.com/css?family=Hind:600|Raleway:500');
-@import 'variables.less';
+@import 'variables.import.less';
 @import '../imports/ui/layouts/project.import.less';
 @import '../imports/ui/layouts/setup.import.less';
 @import '../imports/ui/components/admin/settings.import.less';
 @import '../imports/ui/components/nlu/models/style/NLUModel.import.less';
+@import '../imports/ui/components/nlu/models/style/style.import.less';
+@import '../imports/ui/components/templates/template-upsert/NLUCriteria/NLUCriteriaEditor.import.less';
+@import '../imports/ui/components/utils/style/ImageField.import.less';
+@import '../imports/ui/components/utils/style/style.import.less';
+@import '../imports/ui/components/nlu/activity/style.import.less';
+@import '../imports/ui/components/templates/template-upsert/Template.import.less';
+@import '../imports/ui/components/templates/templates-list/style.import.less';
+@import '../imports/ui/components/common/style.import.less';
+@import '../imports/ui/components/nlu/common/EntityDropdown.import.less';
+@import '../imports/ui/components/nlu/common/IntentDropdown.import.less';
+@import '../imports/ui/components/conversations/style.import.less';
+@import '../imports/ui/components/stories/style.import.less';
+@import '../imports/ui/components/account/style.import.less';
+@import '../imports/ui/components/example_editor/style.import.less';
 
 @sidebar-bg: rgb(45, 47, 74);
 @sidebar-size: 190px;
