+++ botfront/imports/ui/components/nlu/activity/Activity.jsx
@@ -15,7 +15,6 @@ import ActivityDataTable from './ActivityDataTable';
 import ActivityActions from './ActivityActions';
 import { Loading } from '../../utils/Utils';
 import { ActivityCollection } from '../../../../api/activity';
-import './style.less';
 import { NLUModels } from '../../../../api/nlu_model/nlu_model.collection';
 import { getPureIntents } from '../../../../api/nlu_model/nlu_model.utils';
 import { wrapMeteorCallback } from '../../utils/Errors';
