+++ botfront/imports/ui/components/nlu/activity/style.import.less
@@ -1,5 +1,3 @@
-@import '../../../../../client/variables.less';
-
 .confidence-text {
     font-size: 11px;
 }
