+++ botfront/imports/ui/components/nlu/common/EntityDropdown.import.less
@@ -1,5 +1,3 @@
-@import "../../../../../client/variables.less";
-
 .entity-dropdown.ui.selection.dropdown {
   border-color: fade(@blue, 50%);
   border-top: none;
