+++ botfront/imports/ui/components/nlu/common/IntentDropdown.import.less
@@ -1,5 +1,3 @@
-@import "../../../../../client/variables.less";
-
 .intent-dropdown.ui.selection.dropdown {
   border-color: fade(@purple, 50%);
   border-top: none;
