+++ botfront/imports/ui/components/stories/Stories.jsx
@@ -7,7 +7,6 @@ import React from 'react';
 import { wrapMeteorCallback } from '../utils/Errors';
 import ItemsBrowser from '../common/Browser';
 import StoriesEditor from './StoriesEditor';
-import './style.less';
 
 class Stories extends React.Component {
     constructor(props) {
