+++ botfront/imports/ui/components/stories/style.import.less
@@ -74,7 +74,7 @@
   
 .story-editor {
   margin-bottom: 20px;
-  .ui.attached.top {
+  .ui.attached.menu.top {
     border-color: #eeeeee;
     .item {
       padding: 0px 16px;
