+++ botfront/imports/ui/components/templates/template-upsert/NLUCriteria/NLUCriteriaEditor.import.less
@@ -1,4 +1,4 @@
-@import "../../../../../../client/variables.less";
+@import "../../../../../../client/variables.import.less";
 
 // .nlu-criteria-editor .ui.search.selection.dropdown > input.search {
 //   line-height: 1.2142em;
