+++ botfront/package.json
@@ -10,7 +10,7 @@
     "docs:dev": "npx vuepress dev docs",
     "docs:build": "npx vuepress build docs",
     "test-watch": "TEST_BROWSER_DRIVER=chrome meteor test --driver-package meteortesting:mocha --port 3100",
-    "postinstall": "cp -R node_modules/semantic-ui-css/themes/default/assets/fonts ./public/",
+    "postinstall": "cp -R node_modules/semantic-ui-css/themes/default/assets/fonts ./public/ && cp -R node_modules/semantic-ui-css/themes/default/assets/fonts ./client/",
     "licenses:summary": "license-checker --summary",
     "licenses:csv": "license-checker --csv --out botfront-licenses.csv",
     "release": "standard-version",
@@ -85,7 +85,7 @@
     "autoprefixer": "^6.3.1",
     "babel-eslint": "^10.0.1",
     "babel-loader": "^8.0.6",
-    "chromedriver": "^2.42.0",
+    "chromedriver": "^2.46.0",
     "commitizen": "^3.0.7",
     "commitlint": "^7.5.2",
     "core-js": "^2.5.3",
@@ -101,6 +101,7 @@
     "less": "2.5.0",
     "less-loader": "^5.0.0",
     "license-checker": "^25.0.1",
+    "selenium-webdriver": "^3.0.0-beta-2",
     "vuepress": "^1.0.2",
     "webpack": "^4.28.4"
   },
